feat: rename angular modules to follow conventions
This commit renames all Angular module and routing module files to use the conventional dot notation (e.g., app.module.ts). It also updates all import statements to reflect these changes.
