+++ cosqol-dashboard/package.json
@@ -21,6 +21,9 @@
     ]
   },
   "private": true,
+  "engines": {
+    "node": "18.x"
+  },
   "dependencies": {
     "@angular/common": "^20.3.0",
     "@angular/compiler": "^20.3.0",
