+++ cosqol-dashboard/src/app/app.module.ts
@@ -1,7 +1,7 @@
 import { NgModule, provideBrowserGlobalErrorListeners } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
-import { AppRoutingModule } from './app-routing-module';
+import { AppRoutingModule } from './app-routing.module';
 import { App } from './app';
 import { MainLayout } from './layout/main-layout/main-layout';
 import { Sidebar } from './layout/sidebar/sidebar';
