+++ cosqol-dashboard/src/app/features/dashboard/dashboard.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { DashboardRoutingModule } from './dashboard-routing-module';
+import { DashboardRoutingModule } from './dashboard-routing.module';
 import { Overview } from './pages/overview/overview';
