+++ cosqol-dashboard/src/app/features/subscriptions/subscriptions.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { SubscriptionsRoutingModule } from './subscriptions-routing-module';
+import { SubscriptionsRoutingModule } from './subscriptions-routing.module';
 import { Subscriptions } from './subscriptions';
 import { SubscriptionList } from './pages/subscription-list/subscription-list';
 import { SubscriptionTable } from './components/subscription-table/subscription-table';
