+++ cosqol-dashboard/src/app/features/withdraw/withdraw.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { WithdrawRoutingModule } from './withdraw-routing-module';
+import { WithdrawRoutingModule } from './withdraw-routing.module';
 import { Withdraw } from './withdraw';
 import { WithdrawFunds } from './pages/withdraw-funds/withdraw-funds';
 import { WithdrawalForm } from './components/withdrawal-form/withdrawal-form';
