+++ cosqol-dashboard/src/main.ts
@@ -1,5 +1,5 @@
 import { platformBrowser } from '@angular/platform-browser';
-import { AppModule } from './app/app-module';
+import { AppModule } from './app/app.module';
 
 platformBrowser().bootstrapModule(AppModule, {
   ngZoneEventCoalescing: true,
